% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{.member}
\alias{.member}
\title{Returns info about the ith member of the clan}
\usage{
.member(i, x, d = lubridate::now())
}
\arguments{
\item{i}{integer: ith member of the clan}

\item{x:}{an object of class clan}
}
\description{
Returns info about the ith member of the clan
}
\examples{
member ( 4, x = my_clan)
}
