% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{export.clan}
\alias{export.clan}
\title{export method for objects of class clan.
save a clan object in path. File name si ex-tag-today.rds}
\usage{
\method{export}{clan}(x, path)
}
\arguments{
\item{x:}{an object of class clan}
}
\description{
export method for objects of class clan.
save a clan object in path. File name si ex-tag-today.rds
}
\examples{
export (my_clan, '~/')
}
