% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{info.clan}
\alias{info.clan}
\title{Info method for objects of class clan.}
\usage{
\method{info}{clan}(x, d = lubridate::now())
}
\arguments{
\item{x:}{an object of class clan}
}
\description{
Returns a tibble of info about the clan
}
\examples{
info (my_clan)
}
