% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{load_clan}
\alias{load_clan}
\title{load clan data}
\usage{
load_clan(tag, auth_key)
}
\arguments{
\item{tag}{clan tag}

\item{token}{a token}
}
\description{
load clan data
}
\examples{
load_clan( tag = "#282UJRU28", token('~/dev/rcoc/tokens/token001'))
}
