% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{member.clan}
\alias{member.clan}
\title{Member method for objects of class clan.}
\usage{
\method{member}{clan}(x)
}
\arguments{
\item{x:}{an object of class clan}
}
\description{
Returns a tibble of info about all members of the clan
}
\examples{
member (my_clan)
}
