% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{tag.clan}
\alias{tag.clan}
\title{tag method for objects of class clan.
Returns the tag a  clan}
\usage{
\method{tag}{clan}(x)
}
\arguments{
\item{x:}{an object of class clan}
}
\description{
tag method for objects of class clan.
Returns the tag a  clan
}
\examples{
tag (my_clan)
}
